feat: refactor extract_bridge_transfer_id into testing utils file
