+++ protocol-units/bridge/integration-tests/tests/movement_eth.rs
@@ -1,50 +1,27 @@
 use tokio::time::{sleep, Duration}; // Add these imports
 
-use alloy::{
-	hex,
-	node_bindings::Anvil,
-	primitives::{address, keccak256},
-	providers::Provider,
-};
+use alloy::primitives::keccak256;
 use anyhow::Result;
-
-use aptos_sdk::{coin_client::CoinClient, types::LocalAccount};
+use bridge_integration_tests::utils::{self as test_utils};
+use aptos_sdk::coin_client::CoinClient;
 use bridge_integration_tests::TestHarness;
 use bridge_shared::{
-	bridge_contracts::{BridgeContractCounterparty, BridgeContractInitiator, BridgeContractInitiatorError},
+	bridge_contracts::{BridgeContractInitiator, BridgeContractInitiatorError},
 	types::{
 		Amount, AssetType, BridgeTransferId, HashLock, HashLockPreImage, InitiatorAddress,
 		RecipientAddress, TimeLock,
 	},
 };
 
-use ethereum_bridge::types::EthAddress;
-use movement_bridge::utils::{self, MovementAddress};
+use movement_bridge::utils::{self as movement_utils, MovementAddress};
 
-use futures::{
-	channel::mpsc::{self, UnboundedReceiver},
-	StreamExt,
-};
-use rand;
-use tiny_keccak::{Hasher as KeccakHasher, Keccak};
 use tokio::{
-	self,
-	process::{Child, Command},
+	self
 };
 
 use aptos_types::account_address::AccountAddress;
 use tracing::{debug, info};
 use tracing_subscriber;
-
-struct ChildGuard {
-	child: Child,
-}
-
-impl Drop for ChildGuard {
-	fn drop(&mut self) {
-		let _ = self.child.kill();
-	}
-}
 
 #[tokio::test]
 async fn test_movement_client_build_and_fund_accounts() -> Result<(), anyhow::Error> {
@@ -120,11 +97,11 @@ async fn test_movement_client_initiate_and_complete_transfer() -> Result<(), any
                         let mint_amount = 200 * 100_000_000; // Assuming 8 decimals for MovETH
 
                         let mint_args = vec![
-                                utils::serialize_address_initiator(&movement_client.signer().address())?, // Mint to initiator's address
-                                utils::serialize_u64_initiator(&mint_amount)?,                     // Amount to mint (200 MovETH)
+                                movement_utils::serialize_address_initiator(&movement_client.signer().address())?, // Mint to initiator's address
+                                movement_utils::serialize_u64_initiator(&mint_amount)?,                     // Amount to mint (200 MovETH)
                         ];
          
-                        let mint_payload = utils::make_aptos_payload(
+                        let mint_payload = movement_utils::make_aptos_payload(
                                 movement_client.counterparty_address, // Address where moveth module is published
                                 "moveth",
                                 "mint",
@@ -132,7 +109,7 @@ async fn test_movement_client_initiate_and_complete_transfer() -> Result<(), any
                                 mint_args,
                         );
         
-                        utils::send_and_confirm_aptos_transaction(&movement_client.rest_client, movement_client.signer(), mint_payload)
+                        movement_utils::send_and_confirm_aptos_transaction(&movement_client.rest_client, movement_client.signer(), mint_payload)
                                 .await
                                 .map_err(|_| BridgeContractInitiatorError::MintError)?; // New error variant for mint failure
         
@@ -155,70 +132,7 @@ async fn test_movement_client_initiate_and_complete_transfer() -> Result<(), any
 		let sender_address = movement_client.signer().address();
 		let sequence_number = 0; // Replace this with the correct sequence number
 
-		let mut bridge_transfer_id: Option<String> = None;
-
-		let rest_client = movement_client.rest_client();
-		let transactions = rest_client
-			.get_account_transactions(sender_address, Some(sequence_number), Some(20))
-			.await
-			.map_err(|e| anyhow::Error::msg(format!("Failed to get transactions: {:?}", e)))?;
-
-		if let Some(transaction) = transactions.into_inner().last() {
-			if let aptos_sdk::rest_client::Transaction::UserTransaction(user_txn) = transaction {
-				for event in &user_txn.events {
-					// Log the entire event details for debugging
-					info!("Event: {:?}", event);
-
-					if let aptos_sdk::rest_client::aptos_api_types::MoveType::Struct(struct_tag) =
-						&event.typ
-					{
-						if struct_tag.module.as_str() == "atomic_bridge_initiator"
-							&& struct_tag.name.as_str() == "BridgeTransferInitiatedEvent"
-						{
-							bridge_transfer_id = Some(
-								event
-									.data
-									.get("bridge_transfer_id")
-									.ok_or_else(|| {
-										anyhow::Error::msg("bridge_transfer_id not found")
-									})?
-									.as_str()
-									.ok_or_else(|| {
-										anyhow::Error::msg("Invalid bridge_transfer_id format")
-									})?
-									.to_string(),
-							);
-
-							info!("Extracted bridge transfer id: {:?}", bridge_transfer_id);
-							break; // Stop searching after we find the event
-						}
-					}
-				}
-
-				if bridge_transfer_id.is_none() {
-					return Err(anyhow::Error::msg("No matching event found in the transaction"));
-				}
-			} else {
-				return Err(anyhow::Error::msg("Not a user transaction"));
-			}
-		} else {
-			return Err(anyhow::Error::msg(
-				"No transaction found for the provided sequence number",
-			));
-		}
-		// Unwrap the bridge_transfer_id to be used later
-		let bridge_transfer_id = bridge_transfer_id.unwrap();
-
-		let hex_str = bridge_transfer_id.trim_start_matches("0x");
-
-		// Decode the hex string into a Vec<u8>
-		let decoded_vec = hex::decode(hex_str)
-			.map_err(|_| anyhow::Error::msg("Failed to decode hex string into Vec<u8>"))?;
-
-		// Convert the Vec<u8> into a [u8; 32]
-		let bridge_transfer_id: [u8; 32] = decoded_vec
-			.try_into()
-			.map_err(|_| anyhow::Error::msg("Failed to convert decoded Vec<u8> to [u8; 32]"))?;
+		let bridge_transfer_id: [u8; 32] = test_utils::extract_bridge_transfer_id(movement_client, sender_address, sequence_number).await?;
 
 		info!("Bridge transfer id: {:?}", bridge_transfer_id);
 
@@ -314,11 +228,11 @@ async fn test_movement_client_initiate_and_refund_transfer() -> Result<(), anyho
                         let mint_amount = 200 * 100_000_000; // Assuming 8 decimals for MovETH
 
                         let mint_args = vec![
-                                utils::serialize_address_initiator(&movement_client.signer().address())?, // Mint to initiator's address
-                                utils::serialize_u64_initiator(&mint_amount)?,                     // Amount to mint (200 MovETH)
+                                movement_utils::serialize_address_initiator(&movement_client.signer().address())?, // Mint to initiator's address
+                                movement_utils::serialize_u64_initiator(&mint_amount)?,                     // Amount to mint (200 MovETH)
                         ];
          
-                        let mint_payload = utils::make_aptos_payload(
+                        let mint_payload = movement_utils::make_aptos_payload(
                                 movement_client.counterparty_address, // Address where moveth module is published
                                 "moveth",
                                 "mint",
@@ -326,7 +240,7 @@ async fn test_movement_client_initiate_and_refund_transfer() -> Result<(), anyho
                                 mint_args,
                         );
         
-                        utils::send_and_confirm_aptos_transaction(&movement_client.rest_client, movement_client.signer(), mint_payload)
+                        movement_utils::send_and_confirm_aptos_transaction(&movement_client.rest_client, movement_client.signer(), mint_payload)
                                 .await
                                 .map_err(|_| BridgeContractInitiatorError::MintError)?; // New error variant for mint failure
         
@@ -344,79 +258,14 @@ async fn test_movement_client_initiate_and_refund_transfer() -> Result<(), anyho
 				.expect("Failed to initiate bridge transfer");
 		} // End of mutable borrow scope
 
-		// Immutable borrow to extract the bridge transfer ID from the transaction
 		let movement_client = harness.movement_client_mut().expect("Failed to get MovementClient");
 		let sender_address = movement_client.signer().address();
 		let sequence_number = 0; // Replace this with the correct sequence number
 
-		let mut bridge_transfer_id: Option<String> = None;
-
-		let rest_client = movement_client.rest_client();
-		let transactions = rest_client
-			.get_account_transactions(sender_address, Some(sequence_number), Some(20))
-			.await
-			.map_err(|e| anyhow::Error::msg(format!("Failed to get transactions: {:?}", e)))?;
-
-		if let Some(transaction) = transactions.into_inner().last() {
-			if let aptos_sdk::rest_client::Transaction::UserTransaction(user_txn) = transaction {
-				for event in &user_txn.events {
-					// Log the entire event details for debugging
-					info!("Event: {:?}", event);
-
-					if let aptos_sdk::rest_client::aptos_api_types::MoveType::Struct(struct_tag) =
-						&event.typ
-					{
-						if struct_tag.module.as_str() == "atomic_bridge_initiator"
-							&& struct_tag.name.as_str() == "BridgeTransferInitiatedEvent"
-						{
-							bridge_transfer_id = Some(
-								event
-									.data
-									.get("bridge_transfer_id")
-									.ok_or_else(|| {
-										anyhow::Error::msg("bridge_transfer_id not found")
-									})?
-									.as_str()
-									.ok_or_else(|| {
-										anyhow::Error::msg("Invalid bridge_transfer_id format")
-									})?
-									.to_string(),
-							);
-
-							info!("Extracted bridge transfer id: {:?}", bridge_transfer_id);
-							break; // Stop searching after we find the event
-						}
-					}
-				}
-
-				if bridge_transfer_id.is_none() {
-					return Err(anyhow::Error::msg("No matching event found in the transaction"));
-				}
-			} else {
-				return Err(anyhow::Error::msg("Not a user transaction"));
-			}
-		} else {
-			return Err(anyhow::Error::msg(
-				"No transaction found for the provided sequence number",
-			));
-		}
-		// Unwrap the bridge_transfer_id to be used later
-		let bridge_transfer_id = bridge_transfer_id.unwrap();
-
-		let hex_str = bridge_transfer_id.trim_start_matches("0x");
-
-		// Decode the hex string into a Vec<u8>
-		let decoded_vec = hex::decode(hex_str)
-			.map_err(|_| anyhow::Error::msg("Failed to decode hex string into Vec<u8>"))?;
-
-		// Convert the Vec<u8> into a [u8; 32]
-		let bridge_transfer_id: [u8; 32] = decoded_vec
-			.try_into()
-			.map_err(|_| anyhow::Error::msg("Failed to convert decoded Vec<u8> to [u8; 32]"))?;
+		let bridge_transfer_id: [u8; 32] = test_utils::extract_bridge_transfer_id(movement_client, sender_address, sequence_number).await?;
 
 		info!("Bridge transfer id: {:?}", bridge_transfer_id);
 
-		// Now get the transfer details
 		let details = BridgeContractInitiator::get_bridge_transfer_details(
 			movement_client,
 			BridgeTransferId(bridge_transfer_id),
